requirements.txt
fastapi
uvicorn
streamlit
anthropic
openai
supabase
python-dotenv
requests
pandas